+++ src/app/inventory/InventoryItem.m.scss.d.ts
@@ -1,14 +1,19 @@
 // This file is automatically generated.
 // Please do not change this file!
 interface CssExports {
+  'basic': string;
   'borderless': string;
+  'common': string;
   'complete': string;
   'exotic': string;
+  'exoticMasterwork': string;
   'icon': string;
   'iconOverlay': string;
   'icons': string;
+  'legendary': string;
   'masterwork': string;
   'masterworkOverlay': string;
+  'rare': string;
   'searchHidden': string;
   'subclass': string;
   'subclassPathBottom': string;
